ai-module/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.0
python-dotenv>=1.0.0
groq>=0.9.0
